refactor(store): remove list_responses and simplify pagination
Remove list_responses from the ResponseStore trait — OpenAI has
no list endpoint (GET /v1/responses). This also removes the
generic pagination types (ListParams, Order, ResponsePage),
cursor helpers, and the tenant_created index.

ListParams now lives in the OpenAI-specific input_items module,
used only for the /v1/responses/{id}/input_items endpoint.
